Upgrade Kotlin version to 2.1.10
This commit updates the Kotlin version in the project's build configuration from 2.0.21 to 2.1.10. This upgrade incorporates the latest tooling updates, performance enhancements, and bug fixes, including new language features such as guard conditions in `when`, non-local `break` and `continue`, and multi-dollar string interpolation. Also, the commit includes updates to the K2 compiler and improved compatibility with newer Gradle versions.